pyproject.toml
[tool.poetry]
name = "uqlm"
version = "0.0.5.4"
description = ""
authors = ["Dylan Bouchard <dylan.bouchard@cvshealth.com>", "Mohit Singh Chauhan <mohitsingh.chauhan@cvshealth.com>"]
maintainers = [
    "Dylan Bouchard <dbouchard92@gmail.com>", 
    "Mohit Singh Chauhan <mohitcek@gmail.com>",
    "David Skarbrevik <skarbrevik@gmail.com>",
    "Ho-Kyeong Ra <doyajii1@gmail.com>",
    "Viren Bajaj <virenbajaj4@gmail.com>",
    "Zeya Ahmad <zeyahmd@gmail.com>"
]
readme = "README.md"
packages = [
    { include = "uqlm" },
    { include = "uqlm/scorers" },
    { include = "uqlm/judges" },
    { include = "uqlm/black_box" },
    { include = "uqlm/resources" },
    { include = "uqlm/utils" },
]

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
langchain = "^0.3.7"
transformers = "^4.45.2"
scikit-learn = "^1.5.2"
numpy = "^1.26.4"
matplotlib = "^3.9.2"
optuna = "^4.0.0"
bert-score = "^0.3.0"
bleurt = { git = "https://github.com/google-research/bleurt.git", branch = "master" }
pandas = "^2.2.3"
sentence-transformers = "^3.4.0"
tf-keras = "^2.18.0"
datasets = "^3.3.2"


[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.0.1"
ipywidgets = "^8.1.5"
ruff = "0.9.7"
pre-commit = "^4.1.0"
ipykernel = "^6.29.5"

[tool.poetry.group.test]
optional = true

[tool.poetry.group.test.dependencies]
ipykernel = "^6.29.5"
langchain-openai = "^0.3.6"
pytest = "^8.3.5"
langchain-google-vertexai = "^2.0.14"
datasets = "^3.3.2"
pytest-asyncio = "^0.25.3"
pytest-cov = "^6.0.0"

[tool.pytest.ini_options]
markers = [
    "asyncio"
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
